Add initial `pike` payloads module
This change adds a module for Pike smart contract actions, useful for
deserializing JSON payloads into this native struct.

This also adds a `Pike` variant to the `Payload` enum and several tests
for deserializing the various `PikeAction`s and a `PikePayload`.

This also adds the `transact/protocol-sabre` feature to the experimental
`rest-api-resources-batch-tracking` feature. This allows for
Transactions to be built from the REST API payload structs.

Signed-off-by: Shannyn Telander <telander@bitwise.io>