Add helpers for deserializing pike payloads
This change adds helpers for deserializing pike smart contract actions
from JSON. Mainly, this adds `try_from` implementations for the
`CreateAgentAction` and `PikePayload` to be used when deserializing.
The `CreateAgentAction` is deserialized from its associated builder
struct. This allows for any errors related to malformed payloads to
be more descriptive, as it's coming from the `build` method rather than Serde's automatic deserializing.

The `PikePayload` struct is deserialized from an interim struct, which
holds the specific action as a serde_json Value. The `target` field is
then parsed to determine which action the JSON Value should be
deserialized into.

This change also adds tests for deserializing the individual actions and
the `PikePayload` struct.

Signed-off-by: Shannyn Telander <telander@bitwise.io>