Add submitter observer trait and mod
This interface defines the type of component responsible for reporting
and/or recording what happens inside the submitter. For example, an
observer can write the statuses of submitted batches to a database or log
errors that have occured in the submission process. For testing, the
observer can record the result of a submission to be checked against the
expected result.

Signed-off-by: Chris Eckhardt <eckhardt@bitwise.io>