Implement `RunnableGriddleRestApi`
This change adds a method to run the `RunnableGriddleRestApi`.
This also adds an initial struct for the `GriddleRestApi`, which is
created by the `RunnableGriddleRestApi`'s `run` method. This `run`
method is not implemented, as it will only generate a new version of the
running rest api and the create method has not been implemented for
`GriddleRestApi`.

Signed-off-by: Shannyn Telander <telander@bitwise.io>