Add initial `internals` module to build Griddle
This change adds an `internals` module to Griddle to hold structs used
for building, running, and shutting down Griddle.

This change adds an initial struct, `DLTBackend`, used by Griddle to
determine if a service ID is required for client requests. This also
updates Griddle's rest API to use this struct over the Grid SDK
`Endpoint` struct.

This module is behind an experimental `griddle-builder` feature.

Signed-off-by: Shannyn Telander <telander@bitwise.io>