Update `From` method to check response body size
This change updates the method used to convert a client response from
the `ProxyClient` to check the response body contents before attempting
to convert it to another format. This avoids an issue where the upstream
server may return an empty response body. Previously, this `From`
implementation would error if the response body was empty as it still
attempts to convert the bytes, despite being empty.

Signed-off-by: Shannyn Telander <telander@bitwise.io>