Update Griddle example docker-compose
This change updates the Griddle example's docker-compose.yaml file to
handle recent updates made to Grid's Splinter and Actix (and related)
dependencies. This also updates the griddle containers to start
immediately without waiting for the database, as it will not be started
with a database initially. The containers also needed to be updated to
handle receiving requests.

Signed-off-by: Shannyn Telander <telander@bitwise.io>