Allow `extra_unused_lifetimes` clippy warning
This change allows the `extra_unused_lifetimes` clippy lint warning that
is produced from the Diesel macros. This warning will not be fixed in
Diesel 1.4, but a fix for this warning is in a release candidate for
Diesel 2.0.

For more information on this Diesel issue, see here:
https://github.com/diesel-rs/diesel/issues/3182

For more information on this Clippy lint, see here:
https://rust-lang.github.io/rust-clippy/master/index.html#extra_unused_lifetimes

Signed-off-by: Shannyn Telander <telander@bitwise.io>