Avoid adding to an existing string with `format!`
This change updates how strings are constructed in various places. The
update makes it so a new string is not always allocated when appending
it to an existing string. In some cases, it is necessary to use the
`format` macro to ensure output. In most cases, however, the use of
`format` is replaced with `push_str` or more dynamic string creation.

For more information on this Clippy lint, see here:
https://rust-lang.github.io/rust-clippy/master/index.html#format_push_string.

Signed-off-by: Shannyn Telander <telander@bitwise.io>